amazon search work

+++ source/src/com/zns/comicdroid/amazon/AmazonCacheTask.java
@@ -14,10 +14,13 @@ public class AmazonCacheTask extends AsyncTask<String, Void, Void> {
 		//Clean up old cache files
 		File cachePath = new File(params[0]);
 		File[] cacheFiles = cachePath.listFiles();
-		for (File file : cacheFiles) {
-			long diff = new Date().getTime() - file.lastModified();
-			if (diff > Application.CACHE_AMAZONSEARCH_HOURS * 60 * 60 * 1000) {
-				file.delete();
+		if (cacheFiles != null)
+		{
+			for (File file : cacheFiles) {
+				long diff = new Date().getTime() - file.lastModified();
+				if (diff > Application.CACHE_AMAZONSEARCH_HOURS * 60 * 60 * 1000) {
+					file.delete();
+				}
 			}
 		}
 		return null;
